+++ packages/calcite-components/src/components/combobox-item/resources.ts
@@ -4,17 +4,15 @@ export const CSS = {
   active: "label--active",
   centerContent: "center-content",
   container: "container",
-  custom: "icon--custom",
+  iconCustom: "icon--custom",
   description: "description",
   icon: "icon",
-  iconSelected: "icon--selected",
   label: "label",
   scale: (scale: Scale) => `scale--${scale}` as const,
-  selected: "label--selected",
   shortText: "short-text",
   single: "label--single",
   textContainer: "text-container",
-  title: "title",
+  heading: "heading",
 };
 
 export const SLOTS = {