+++ packages/calcite-components/src/components/combobox/combobox.tsx
@@ -100,6 +100,8 @@ export class Combobox
 
   // #region Private Properties
 
+  private closeButtonEl = createRef<HTMLButtonElement>();
+
   private allSelectedIndicatorChipEl: Chip["el"];
 
   private chipContainerEl: HTMLDivElement;
@@ -1418,8 +1420,7 @@ export class Combobox
     const { activeChipIndex, readOnly, scale, selectionMode, messages } = this;
     return this.selectedItems.map((item, i) => {
       const chipClasses = {
-        chip: true,
-        "chip--active": activeChipIndex === i,
+        [CSS.chip]: true,
       };
       const ancestors = [...getItemAncestors(item)].reverse();
       const itemLabel = getLabel(item);
@@ -1739,9 +1740,9 @@ export class Combobox
         <div
           ariaLive="polite"
           class={{
-            wrapper: true,
-            "wrapper--single": singleSelectionMode || !this.selectedItems.length,
-            "wrapper--active": open,
+            [CSS.wrapper]: true,
+            [CSS.wrapperSingle]: singleSelectionMode || !this.selectedItems.length,
+            [CSS.wrapperActive]: open,
           }}
           onClick={this.clickHandler}
           onKeyDown={this.keyDownHandler}
@@ -1779,6 +1780,7 @@ export class Combobox
               disabled={this.disabled}
               key="close-button"
               label={this.messages.clear}
+              ref={this.closeButtonEl}
               scale={this.scale}
             />
           ) : null}