+++ packages/calcite-components/src/components/tab-title/tab-title.tsx
@@ -1,5 +1,5 @@
 // @ts-strict-ignore
-import { PropertyValues } from "lit";
+import { PropertyValues, isServer } from "lit";
 import { createRef } from "lit-html/directives/ref.js";
 import {
   LitElement,
@@ -24,11 +24,11 @@ import { TabChangeEventDetail, TabCloseEventDetail } from "../tab/interfaces";
 import { TabID, TabLayout, TabPosition } from "../tabs/interfaces";
 import { getIconScale } from "../../utils/component";
 import { IconNameOrString } from "../icon/interfaces";
-import { isBrowser } from "../../utils/browser";
+import { XButton } from "../functional/XButton";
 import { useT9n } from "../../controllers/useT9n";
 import type { Tabs } from "../tabs/tabs";
 import T9nStrings from "./assets/t9n/messages.en.json";
-import { CSS, ICONS } from "./resources";
+import { CSS } from "./resources";
 import { styles } from "./tab-title.scss";
 
 declare global {
@@ -156,7 +156,7 @@ export class TabTitle extends LitElement implements InteractiveComponent {
     if (this.disabled || this.closed) {
       return;
     }
-    const payload = { tab: this.tab };
+    const payload = { tab: this.tab, userTriggered: userTriggered };
     this.calciteInternalTabsActivate.emit(payload);
 
     if (userTriggered) {
@@ -257,7 +257,7 @@ export class TabTitle extends LitElement implements InteractiveComponent {
   }
 
   async load(): Promise<void> {
-    if (isBrowser()) {
+    if (!isServer) {
       this.updateHasText();
     }
     if (this.tab && this.selected) {
@@ -452,18 +452,17 @@ export class TabTitle extends LitElement implements InteractiveComponent {
     const { closable, messages } = this;
 
     return closable ? (
-      <button
-        ariaLabel={messages.close}
-        class={CSS.closeButton}
+      <XButton
         disabled={false}
-        key={CSS.closeButton}
+        focusable={true}
+        key="close-button"
+        label={messages.close}
         onClick={this.closeClickHandler}
         ref={this.closeButtonEl}
+        round={false}
+        scale={this.scale}
         title={messages.close}
-        type="button"
-      >
-        <calcite-icon icon={ICONS.close} scale={getIconScale(this.scale)} />
-      </button>
+      />
     ) : null;
   }