+++ packages/calcite-components/src/components/tab/tab.e2e.ts
@@ -1,5 +1,6 @@
 import { describe } from "vitest";
-import { defaults, renders, hidden } from "../../tests/commonTests";
+import { defaults, renders, hidden, themed } from "../../tests/commonTests";
+import { CSS } from "./resources";
 
 describe("calcite-tab", () => {
   const tabHtml = "<calcite-tab>A tab</calcite-tab>";
@@ -20,5 +21,16 @@ describe("calcite-tab", () => {
       { propertyName: "selected", defaultValue: false },
       { propertyName: "scale", defaultValue: "m" },
     ]);
+  });
+
+  describe("theme", () => {
+    describe("default", () => {
+      themed("calcite-tab", {
+        "--calcite-tab-content-space-y": {
+          shadowSelector: `.${CSS.content}`,
+          targetProp: "paddingBlock",
+        },
+      });
+    });
   });
 });